Changed configuration::ApplicationShouldStart to be more explicit.
Context:
The ApplicationShouldStart function was not considering the value of
application->autostart(), which was not expected considering its name.
These issues caused crashes in downstream code due to unexpected
behavior.

Change:
Change ApplicationShouldStart and GetApplication to allow filtering
based on the value of the application's `autostart` value.
Reworked the GetApplicationsContainingSubstring and Autostart enum for
clarity. This has no present impact because nothing was using
`Autostart::kNo`, but will help to avoid future confusion.

Change-Id: I45c9df8a8d88e25acec53342b9520a92aab72b03
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>