Check that applications are sorted in aos::configuration::GetApplication
Context:
Some unit tests were using GetApplication with an unsorted aos
configuration, which caused undefined behavior because GetApplication
uses std::lower_bound, which requires the input to be sorted to be
useful.

Change:
Add a check that only runs during debug mode. This will ensure that
all tests that use GetApplication are using a sorted configuration,
because we always run the tests in debug mode (among other modes).

Change-Id: I92494948f57b96c07c59617cf3ce488a5afcb236
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>