(aos): Fix send-time race condition in lockless queue
This change modifies the send-time calculation on the shared-memory
queues such that the send time for a message is calculated by whatever
process first observes the message after it is sent. Normally, this will
be the sending process itself, but for situations where a reader manages
to observe the message before the sender finishes timestamping things
the reader will end up populating the timestamps.

The bulk of this change is constructing adequate test cases for this. We
also continue to support the old send-time population behavior in order
to support systems which do not have 64-bit atomics available with which
to allow concurrent writing of 64-bit timestamps.

This also has the inadvertent benefit of turning on more of the lockless
queue tests in our regular CI because it moves the extremely
long-running piece of the lockless_queue_test into a separate test.

Note: From an ad-hoc measurement, this had an ~14% time-penalty on the
lockless_queue_wrap_test, which should pretty tightly correspond to the
time penalty associated with doing nothing except sending + reading on a
single channel.

See https://github.com/RealtimeRoboticsGroup/aos/issues/3

Increments the IPC version because while this does not change the
in-memory layout it *does* change the protocol associated with said
memory.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I9b4fa2d491b9e14279dfb57880aa5facae3f67ff