Add a Bazel macro to compile flatbuffers for Python
The `flatbuffer_py_library` macro collects transitive dependencies,
invokes flatc to generate Python modules for all the objects defined in
the schema and its includes, and defines a `py_library` which contains
them within a top-level module with the provided name.

This top-level module needs to be included in the import paths in the
generated Python code, so we need a minor change to the upstream
flatbuffers codegen code which lets us specify this through a flag.

Note that deps must be targets of the form `//path/to:<name>_srcs`, that
are implicitly created for each `flatbuffer_py_library` target <name>.

The generated output from flatc is different enough for Python that it
warrants its own separate Bazel rule. See the flatbuffer_py_library
documentation for details.

Change-Id: Id0cf96c0e6cc00f8b6b6179b0378f5a2d2d1a6d8
Signed-off-by: Sanjay Narayanan <sanjaynarayanan94@gmail.com>