Create Python bindings for AOS ShmEventLoop
This adds a C API for the AOS EventLoop that Python binds to using cffi.
The Python side of things is designed around the async-await pattern,
largely inspired by the Rust implementation. An EventLoopRuntime serves
as the main interface for Python programs to await events and interact
with the event loop.

The ping_py and pong_py applications demonstrate basic usage of this
API and paradigm.

Change-Id: I0dbec117176b7d29d82d0cdce2538ade8658078c
Signed-off-by: Sanjay Narayanan <sanjaynarayanan94@gmail.com>