Make dependencies available from external repos through repositories.bzl
Lets make it a lot easier for external dependencies to depend on us.
The rawrtc library being vendored in makes it pretty tough.  Instead,
use it from a http_archive.  We aren't changing it enough to matter.

Also, make repositories.bzl which has the common load statements to pull
in elsewhere.

Change-Id: I5d9398aa8da46aa2aa279a838a39068f1a5a48a2
Signed-off-by: Austin Schuh <austin.linux@gmail.com>