Explode on messages sent before time starts in AOS
The log reader gets pretty pissed if you try to read a log with a
message from before time=0 on a node.  Practically speaking, the
real hardware also can't start with negative monotonic times.

Just explode if anyone tries to do it.  This needs some tests, and I'm
sure something is relying on this behavior accidentally.

Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: Ifef7f3124e2c5ff9ecd011b739102480b88a6efe