aos: Add a formalized threading construct in the event loop
This patch adds a new `threads` field in the `Application` table. This
lets users add configuration for threads. The threads can then be
configured at runtime with the new
`EventLoop::ConfigureThreadAndWaitForRun` function.

The goal here is to allow applications to offload work in event
handlers in a way that is configurable via the AOS configuration.

See the documentation in `documentation/aos/docs/threading.md` for
more information.

One note is that I chose not to enhance the
`aos/events/event_loop_param_test.cc` library. This is because writing
death tests was proving untenable. Death tests in gtest use a clone of
the running process to run the code under test. This means that the
code under test will run under a different PID compared to when the
corresponding event loop was created. That means a death test cannot
validate the behaviour of calling the various helper functions from
specific contexts. To work around this, I created a new tool under
`aos/events/event_loop_thread_tester.cc`. This tool is orchestrated
via `aos/events/event_loop_thread_test.py`. The expected behaviour of
the `ShmEventLoop` and `SimulatedEventLoopFactory` are tested there.

There are some alternatives brought up by other folks during the
design discussion of this.

Coroutines: These are not on their own capable of offloading work.
They are useful for interleaving work within a thread that is very
ergonomic for programmers. There is, however, no inherent parallelism
available here.

Processes: We could offload the work to other processes instead of
threads. Threads create risk via data races. Processes, on the other
hand, have much better control over which memory they can read from or
write to. However, communicating between processes is also a lot more
involved. It is possible that someone comes up with a way to make this
ergonomic, but at the moment it is not ergonomic.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I756aeff8ab2c8facef1e044e58e8621820452f38