aos: Fix some issues from the event loop threading patch
During the review for the "Add a formalized threading construct in the
event loop" patch, Brian Griglak pointed out that the main thread in
the tests could be realtime. There's no downside to this so this patch
fixes that issue.

Separately, Payton Rehl pointed out a typo that I missed. This patch
fixes that as well.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I8d0f57f1ef8bed3dda319cdf5fefe159c0e623ad