aos: Fix asan/msan issues with {shm,simulated}_event_loop_thread_test
The tests are failing under the sanitizers. One of the unit tests
relies on the malloc hook detecting a `LOG(INFO)` in the realtime
path. The malloc hooks are not present under these sanitizers,
however.

This patch works around the issue by explicitly using a `LOG(FATAL)`
when sanitizers are in use.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I642d82ce5f04f4ec3d2cc4835d24f28ff8dca1ef