aos: Add more information about threads starting up
I ran into one test where I didn't run the event loop. That resulted
in a hang and was annoying to debug. This patch should make it a
little easier by explicitly telling the user that we're waiting for
the event loop to start running.

I changed the logic in the `SimulatedThreadHandle` a bit. The thread
is now marked realtime at the same time that the `ShmThreadHandle`
does this action. I.e. after waiting for the event loop's ready
signal.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I2d2fb5f2a5bc57e54b750861f787cba784698d93