Explicitly set the CA bundle location for libcurl
We were seeing some errors during reading S3 files.

During S3 log reading, libcurl was complaining with error code 77 indicating a problem with
reading the CA bundle. The direct cause seems to be that the `curl.BUILD` logic for determining where
libcurl should look for the CA bundle is based on the build host, not the runtime environment.
Therefore, if we build on redhat but run on debian, libcurl will look in the wrong spot, resulting
in the error above.

`curl.BUILD` actually provides us with a solution to this - instead of guessing at the bundle location
based on the build host, we can explicitly specify what flavor of linux we are running via a build define.

In this change, we change the define to a flag which is considered better practice in bazel, and we default
this flag to debian in our build. All of our runtime environments use the debian standard for the CA bundle
which is `/etc/ssl/certs/ca-certificates.crt`.

I did also try setting `CURL_CA_BUNDLE` and `AWS_CA_BUNDLE` to override whatever libcurl was compiled with
but that had no effect as far as I could tell.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I1b67111ee17218b41489b228b62600b78e5d7311