AOS: Consistently handle insufficient memory errors in FromFlatbuffer()
I had had a `ABSL_CHECK()` to ensure that the `add_field()` in
`FromFlatbuffer()` succeeded. I originally did this because it is
supposed to always succeed for submessages (because the memory for
submessages is currently always statically preallocated). However, the
same CHECK was also running when `add_field()` was called for an
unspecified-length vector, and in that case `add_field()` can return
`nullptr`. Just make it so that we unconditionally return false (to
indicate insufficient memory) when `add_field()` returns nullptr.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I82603eb100f1ff54b695abd9f88d11ef7c02d8b2