Handle SCTP invalid association ID
In SetStreamPriority method we encounter an EINVAL error when trying to
set stream priorities,
and then when we try to get the association status to debug the issue,
it fails because the
association ID is no longer valid. This change replaces the crash with a
boolean return
that indicates where SetStreamPriority is a success/failure.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I0c8768016aa33a41b30bc4cce967d076e072dbe4