Support building AOS from an external repo
We were assuming that the libpthread.so and libc.so linker scripts were
in a well defined spot relative to the root of the repo.  That doesn't
hold if aos is an external repo.  So, instead, use more dynamic paths
which work when the toolchain isn't in the main repo.

Change-Id: I6d9f9c5480009330a4d15156e622c52893d1f10f
Signed-off-by: Austin Schuh <austin.linux@gmail.com>