Add //tools/platforms/python:has_support and wrapper macros
bzlmod is much more picky about toolchains than WORKSPACE files.  This
shows up with python not having a toolchain for the roboRIO, which then
results in build failures.  Solve that the same way we handle rust + go,
add an explicit support constraint value, and then use macros which
require it.

Change-Id: I1d28a27b49c772decdfe6ad72de59108afe81a77
Signed-off-by: Austin Schuh <austin.linux@gmail.com>