Enable running sctp_test in kubernetes
The test wants to run with admin permission.  What actually matters here
is that sctp_auth is enabled.  Check that before trying to enable it, so
we can have a kubernetes worker with the right permissions already, and
the test will just use them rather than trying to set it up.  This plays
a lot better with Kubernetes.

Change-Id: Ib7d79a6317af3c5f39b47959e8cb3660ca69290b
Signed-off-by: Austin Schuh <austin.linux@gmail.com>